Consolidate metadata size
Fixed block_size_min to be sizeof(block_header_t) minus
the size of metadata instead of the size of the prev_phys_block pointer.
These happen to be of the same size, but it is conceptually wrong.
For example, we may want to expand the metadata per allocation

For convenience, also changed offset_to_block() to take a block size
instead of block size - block_header_overhead

Signed-off-by: Jim Huang <jserv.tw@gmail.com>